+++ provisioner/echo/serve_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/coder/coder/v2/codersdk/drpc"
 	"github.com/coder/coder/v2/provisioner/echo"
 	"github.com/coder/coder/v2/provisionersdk"
 	"github.com/coder/coder/v2/provisionersdk/proto"
@@ -19,7 +20,7 @@ func TestEcho(t *testing.T) {
 	workdir := t.TempDir()
 
 	// Create an in-memory provisioner to communicate with.
-	client, server := provisionersdk.MemTransportPipe()
+	client, server := drpc.MemTransportPipe()
 	ctx, cancelFunc := context.WithCancel(context.Background())
 	t.Cleanup(func() {
 		_ = client.Close()
