+++ provisionersdk/serve_test.go
@@ -10,6 +10,7 @@ import (
 	"go.uber.org/goleak"
 	"storj.io/drpc/drpcconn"
 
+	"github.com/coder/coder/v2/codersdk/drpc"
 	"github.com/coder/coder/v2/provisionersdk"
 	"github.com/coder/coder/v2/provisionersdk/proto"
 	"github.com/coder/coder/v2/testutil"
@@ -23,7 +24,7 @@ func TestProvisionerSDK(t *testing.T) {
 	t.Parallel()
 	t.Run("ServeListener", func(t *testing.T) {
 		t.Parallel()
-		client, server := provisionersdk.MemTransportPipe()
+		client, server := drpc.MemTransportPipe()
 		defer client.Close()
 		defer server.Close()
 
@@ -65,7 +66,7 @@ func TestProvisionerSDK(t *testing.T) {
 
 	t.Run("ServeClosedPipe", func(t *testing.T) {
 		t.Parallel()
-		client, server := provisionersdk.MemTransportPipe()
+		client, server := drpc.MemTransportPipe()
 		_ = client.Close()
 		_ = server.Close()
